pyproject.toml
[project]
name = "szamlazz-collections-mcp"
version = "0.1.0"
description = "FastMCP server for Szamlazz.hu collections automation"
authors = [{name = "Automated Agent"}]
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.1.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "jinja2>=3.1.4",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.3",
    "pytest>=8.2.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
src = ["src", "tests"]
select = ["E", "F", "I", "B", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
